back-end/requirements.txt
flask
openai
flask-cors
python-dotenv
